requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.9.1
asgiref==3.4.1
astroid==2.8.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bleach==4.1.0
boto3==1.19.2
botocore==1.22.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
cssselect==1.1.0
decorator==5.1.0
dicttoxml==1.7.4
distlib==0.3.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.7
django-anymail==8.5
django-binary-database-files==1.0.15
django-colorfield==0.4.3
django-crispy-forms==1.14.0
django-formtools==2.3
django-json-tag==0.0.1
django-otp==1.1.3
django-phonenumber-field==6.1.0
django-storages==1.12.2
django-two-factor-auth==1.13.2
docutils==0.17.1
filelock==3.1.0
future==0.18.2
geocoder==1.38.1
geoip2==4.4.0
gunicorn==20.1.0
h11==0.12.0
idna==3.2
importlib-metadata==4.8.1
ip2geotools==0.1.6
IP2Location==8.6.2
isort==5.9.3
jeepney==0.8.0
jmespath==0.10.0
keyring==23.2.1
lazy-object-proxy==1.6.0
lxml==4.6.3
maxminddb==2.2.0
mccabe==0.6.1
multidict==5.2.0
outcome==1.1.0
packaging==21.0
password-validator==1.0
phonenumbers==8.12.46
pip-review==1.1.0
pkginfo==1.7.1
platformdirs==2.4.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
pylint==2.11.1
pyparsing==2.4.7
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
qrcode==7.3.1
ratelim==0.1.6
readme-renderer==30.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
s3transfer==0.5.0
SecretStorage==3.3.2
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
twine==3.4.2
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
webencodings==0.5.1
wrapt==1.13.2
wsproto==1.0.0
yarl==1.7.0
zipp==3.6.0
